refactor: consolidate color utilities and enhance drawing helpers in C++ examples
